# nursery-inventory-system/nursery-inventory-system/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plant-batches:
    get:
      operationId: listPlantBatches
      tags: [plant-batches]
      summary: List all plant batches
      parameters:
        - in: query
          name: search
          schema: { type: string }
          description: Search across batchNumber, supplier, plant name
        - in: query
          name: plantId
          schema: { type: integer }
        - in: query
          name: nurseryLocation
          schema: { type: string }
        - in: query
          name: supplier
          schema: { type: string }
      responses:
        "200":
          description: List of batches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlantBatch"

    post:
      operationId: createPlantBatch
      tags: [plant-batches]
      summary: Create a plant batch (purchase)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantBatchInput"
      responses:
        "201":
          description: Created batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantBatch"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Batch number already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /plant-batches/{id}:
    get:
      operationId: getPlantBatch
      tags: [plant-batches]
      summary: Get a plant batch
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Batch record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantBatch"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

    put:
      operationId: updatePlantBatch
      tags: [plant-batches]
      summary: Update a plant batch
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantBatchUpdate"
      responses:
        "200":
          description: Updated batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantBatch"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

    delete:
      operationId: deletePlantBatch
      tags: [plant-batches]
      summary: Delete a plant batch
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Production Batches ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Plant Master schemas ─────────────────────────────────────────────────
...
    PlantBatchPlant:
      type: object
      required: [id, plantCode, commonName, botanicalName, potSize]
      properties:
        id:           { type: integer }
        plantCode:    { type: string }
        commonName:   { type: string }
        botanicalName: { type: string }
        potSize:      { type: string }
        imageUrl:     { type: ["string", "null"] }
    PlantBatch:
      type: object
      required:
        - id
        - batchNumber
        - supplier
        - plantId
        - plant
        - potSize
        - quantityPurchased
        - purchasePricePerPlant
        - transportationCost
        - otherCosts
        - totalBatchCost
        - costPerPlant
        - purchaseDate
        - currentQuantity
        - nurseryLocation
        - createdAt
        - updatedAt
      properties:
        id:                    { type: integer }
        batchNumber:           { type: string }
        supplier:              { type: string }
        plantId:               { type: integer }
        plant:
          $ref: "#/components/schemas/PlantBatchPlant"
        potSize:               { type: string }
        quantityPurchased:     { type: integer }
        purchasePricePerPlant: { type: number }
        transportationCost:    { type: number }
        otherCosts:            { type: number }
        totalBatchCost:        { type: number }
        costPerPlant:          { type: number }
        purchaseDate:          { type: string, format: date }
        currentQuantity:       { type: integer }
        nurseryLocation:       { type: string }
        createdAt:             { type: string, format: date-time }
        updatedAt:             { type: string, format: date-time }
    PlantBatchInput:
      type: object
      required:
        - supplier
        - plantId
        - potSize
        - quantityPurchased
        - purchasePricePerPlant
        - purchaseDate
        - nurseryLocation
      properties:
        batchNumber:           { type: string }
        supplier:              { type: string, minLength: 1 }
        plantId:               { type: integer }
        potSize:               { type: string, minLength: 1 }
        quantityPurchased:     { type: integer, minimum: 1 }
        purchasePricePerPlant: { type: number, minimum: 0 }
        transportationCost:    { type: number, minimum: 0 }
        otherCosts:            { type: number, minimum: 0 }
        purchaseDate:          { type: string, format: date }
        nurseryLocation:       { type: string, minLength: 1 }
    PlantBatchUpdate:
      type: object
      properties:
        batchNumber:           { type: string, minLength: 1 }
        supplier:              { type: string, minLength: 1 }
        plantId:               { type: integer }
        potSize:               { type: string, minLength: 1 }
        quantityPurchased:     { type: integer, minimum: 1 }
        purchasePricePerPlant: { type: number, minimum: 0 }
        transportationCost:    { type: number, minimum: 0 }
        otherCosts:            { type: number, minimum: 0 }
        purchaseDate:          { type: string, format: date }
        currentQuantity:       { type: integer, minimum: 0 }
        nurseryLocation:       { type: string, minLength: 1 }

    # ── Production Batch schemas ────────────────────────────────────────────
...